// Frontend/todo-app/src/clients/todo-api-client.ts
export class TodoItemDto implements ITodoItemDto {
    id!: string;
    title!: string;
    checked!: boolean;
    todoListId!: string;

    [key: string]: any;

    constructor(data?: ITodoItemDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.id = _data["id"];
            this.title = _data["title"];
            this.checked = _data["checked"];
            this.todoListId = _data["todoListId"];
        }
    }

    static fromJS(data: any): TodoItemDto {
        data = typeof data === 'object' ? data : {};
        let result = new TodoItemDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["id"] = this.id;
        data["title"] = this.title;
        data["checked"] = this.checked;
        data["todoListId"] = this.todoListId;
        return data;
    }
}

export interface ITodoItemDto {
    id: string;
    title: string;
    checked: boolean;
    todoListId: string;

    [key: string]: any;
}

export class TodoItemUpdateDto implements ITodoItemUpdateDto {
    title!: string;
    checked!: boolean;

    [key: string]: any;

    constructor(data?: ITodoItemUpdateDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.title = _data["title"];
            this.checked = _data["checked"];
        }
    }

    static fromJS(data: any): TodoItemUpdateDto {
        data = typeof data === 'object' ? data : {};
        let result = new TodoItemUpdateDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["title"] = this.title;
        data["checked"] = this.checked;
        return data;
    }
}

export interface ITodoItemUpdateDto {
    title: string;
    checked: boolean;

    [key: string]: any;
}
// ...
export class TodoListDto implements ITodoListDto {
    id!: string;
    name!: string;
    todoItems!: TodoItemDto[];

    [key: string]: any;

    constructor(data?: ITodoListDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.todoItems = [];
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.id = _data["id"];
            this.name = _data["name"];
            if (Array.isArray(_data["todoItems"])) {
                this.todoItems = [] as any;
                for (let item of _data["todoItems"])
                    this.todoItems!.push(TodoItemDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): TodoListDto {
        data = typeof data === 'object' ? data : {};
        let result = new TodoListDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["id"] = this.id;
        data["name"] = this.name;
        if (Array.isArray(this.todoItems)) {
            data["todoItems"] = [];
            for (let item of this.todoItems)
                data["todoItems"].push(item ? item.toJSON() : undefined as any);
        }
        return data;
    }
}

export interface ITodoListDto {
    id: string;
    name: string;
    todoItems: TodoItemDto[];

    [key: string]: any;
}